s3air

+++ Oxygen/lemonscript/source/lemon/runtime/RuntimeFunction.cpp
@@ -183,7 +183,7 @@ namespace lemon
 					// Take a shortcut by skipping the jump opcode and directly pointing to its target as next opcode
 					//  -> But only do that for jumps forward, otherwise it's possible that script execution can get stuck in an infinite loop
 					//  -> That's because counted steps are only checked in actually executed jumps, but not in those that we optimize away here
-					RuntimeOpcode* targetPointer = reinterpret_cast<RuntimeOpcode*>(runtimeOpcode.mNext->getParameter<uint64>());
+					RuntimeOpcode* targetPointer = reinterpret_cast<RuntimeOpcode*>((size_t)runtimeOpcode.mNext->getParameter<uint64>());
 					RuntimeOpcode* ownPointer = &runtimeOpcode;
 					if (targetPointer <= ownPointer)
 						break;
